modifications de serverconnection pour le téléchargement d'un thème, et de JarLevel pour l'enregistrement d'un niveau MY

+++ Winds/src/addon/JarLevel.java
@@ -127,7 +127,7 @@ public class JarLevel {
 		
 		if( !levelsFolder.canWrite() )
 			throw new Exception("No writing access rights onto levels folder");
-		if(json.length() >= (int) levelsFolder.getFreeSpace())
+		if((long)json.length() >= levelsFolder.getFreeSpace())
 			throw new Exception("Not enough free space into HDD");
 	}
 	/**
